requirements.txt
# Зависимости проекта
aiogram==3.4.1
openai==1.12.0
python-docx==1.1.0
python-dotenv==1.0.0
asyncio
aiofiles==23.2.1
